+++ cos_sin.py
@@ -1,0 +1,6 @@
+import math
+
+# Посчитайте значение выражения: cos(x)*sin(x)
+
+x = int(input('insert integer:'))
+print('cosinus of x is: ', math.cos(x), '\nsinus of x is: ', math.sin(x))
