+++ cycle_power_integer.py
@@ -1,0 +1,6 @@
+# Возвести число в степень с помощью цикла
+
+value = int(input('insert any int:'))
+while 0 < value < 100000:
+    value = value*value
+    print(value*value)
