+++ number_squares.py
@@ -1,0 +1,4 @@
+# Вывести квадраты всех чисел от 1 до 10
+
+for i in range(1, 11):
+    print(i*i)
