+++ parsing_of_line_with_digits.py
@@ -1,0 +1,7 @@
+# Дано натуральное число N. Выведите все его цифры по одной
+# в обратном порядке, разделяя их пробелами или новыми строками.1/5
+
+list1 = []
+num = 3456789
+for digit in reversed(str(num)):
+    print(digit, end=" ")
