+++ print_current_time.py
@@ -1,0 +1,5 @@
+import time
+
+# Выведите текущее время в любом формате
+
+print(time.strftime('%H:%M:%S', (time.localtime())))
