+++ print_positive_even_number.py
@@ -1,0 +1,7 @@
+# Напишите программу используя цикл с предусловием while для вывода каждого четного положительного числа от 0 до 20.
+
+a = 0
+while a in range(0, 20):
+    a = a+1
+    if a % 2 == 0:
+        print(a)
