+++ print_random_values.py
@@ -1,0 +1,5 @@
+import random
+
+# Выведите случайные числа в заданном пользователем диапазоне
+
+print(random.randint(3, 30))
