+++ print_uppercase_letters.py
@@ -1,0 +1,6 @@
+# Вывести все заглавные буквы в строке
+
+a = 'This is Hard timE fOr ouR countrY'
+for i in a:
+    if i.isupper():
+        print(i, end='')
