+++ squares_until_hundred.py
@@ -1,0 +1,8 @@
+# Выводить квадраты чисел до тех пор пока они меньше 1000
+
+a = 0
+while a <= 1000:
+    a += 1
+    i = a*a
+    if i < 1000:
+        print(i)
