+++ sum_of_input_numbers.py
@@ -1,0 +1,6 @@
+# Напишите программу, которая выводит сумму введенных пользователем чисел. Числа вводятся одной строкой.
+
+a, b = input('print two values separated by comma: ').split(',')
+a = int(a)
+b = int(b)
+print(a+b)
